refactor: rename publication feature to publications
