+++ lib/app/main_app.dart
@@ -3,12 +3,12 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../core/extensions/l10n_extension.dart';
 import '../core/shared/cubits/locale_cubit/locale_cubit.dart';
-import '../core/shared/cubits/theme_cubit.dart';
+import '../core/shared/cubits/theme_cubit/theme_cubit.dart';
 import '../core/theme/app_theme.dart';
 import '../features/auth/presentation/bloc/auth_bloc.dart';
 import '../features/comment/presentation/bloc/comment_bloc.dart';
 import '../features/profile/presentation/bloc/profile_bloc.dart';
-import '../features/publication/presentation/bloc/publication_bloc.dart';
+import '../features/publications/presentation/bloc/publication_bloc.dart';
 import '../l10n/generated/app_localizations.dart';
 import 'router/app_router.dart';
 import 'service_locator/service_locator.dart';
