+++ lib/app/router/app_router.dart
@@ -4,8 +4,8 @@ import '../../core/shared/pages/app_view.dart';
 import '../../core/shared/pages/error_page.dart';
 import '../../features/auth/presentation/pages/login_page.dart';
 import '../../features/profile/presentation/pages/profil_page.dart';
-import '../../features/publication/presentation/pages/add_publication_page.dart';
-import '../../features/publication/presentation/pages/publication_list_page.dart';
+import '../../features/publications/presentation/pages/add_publication_page.dart';
+import '../../features/publications/presentation/pages/publication_list_page.dart';
 import '../../features/recommendations/presentation/pages/recommendations_page.dart';
 import 'app_routes.dart';
