+++ lib/features/publications/data/models/publication_model.dart
