+++ lib/features/publications/di/publication_injection.dart
