+++ lib/features/publications/domain/entities/publication.dart
