+++ lib/features/publications/domain/entities/publication_type.dart
