+++ lib/features/publications/presentation/widgets/autoplay_video.dart
